requirements.txt
pynput>=1.7.6
Pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy>=1.26.0
PySide6>=6.6.0
